parser/wiki/requirements.txt
certifip==2024.8.30
crc64iso==0.0.2
mediawiki_dump==1.3.0
mwparserfromhell==0.6
pillow==11.0.0
pycurl==7.45.3
regex==2024.11.6
